feat: add explore-section
