+++ museum/src/index.html
@@ -164,7 +164,20 @@
             </section>
             <section class="section section-black">
                 <div class="section-container">
-                    
+                    <article class="explore-text-content white-text">
+                        <h2 class="section-title explore-title underline-title">Picture explore</h2>
+                        <p class="explore-text-info">Las Meninas is a 1656 painting by Diego Velázquez, the leading <br> artist of the Spanish Golden Age.</p>
+                        <p class="explore-text-info">It was cleaned in 1984 to remove a <span class="gold-text">"yellow veil"</span> of dust that had gathered since the previous restoration in the 19th century.</p>
+                        <p class="explore-text-info">The cleaning provoked furious protests, not because the picture had been damaged in any way, but because it looked different.</p>
+                    </article>
+                    <div class="img-explore-container">
+                        <div class="img-explore-img">
+                          <img src="./assets/img/explore-slider/after.jpg">
+                        </div>
+                        <div class="img-explore-img img-explore-overlay">
+                          <img src="./assets/img/explore-slider/before.jpg">
+                        </div>
+                    </div>
                 </div>
             </section>
             <section class="section">
